geo_agent/pyproject.toml
[project]
name = "geo-agent"
version = "0.1.0"
description = "OpenAI compatible agent service powered by qwen3-max"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "dashscope>=1.18.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "structlog>=24.1.0",
    "pyyaml>=6.0.1",
    "httpx>=0.26.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
include = [
    "app",
    "main.py",
    "config.yaml",
]

[tool.uv]
# 使用项目本地虚拟环境
managed = true

dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "openai>=1.10.0",
]
